requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
EasyProcess==1.1
entrypoint2==1.1
flask==2.3.3
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonalias==0.1.1
MarkupSafe==2.1.3
mss==9.0.1
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
outcome==1.2.0
packaging==23.1
Pillow==10.0.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypng==0.20220715.0
PyQRCode==1.2.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dotenv==1.0.0
pytweening==1.0.7
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rubicon-objc==0.4.6
segno==1.5.2
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
solders==0.18.1
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
tweepy==4.14.0
tweetshot==0.0.1
typing_extensions==4.8.0
urllib3==2.0.5
webdriver-manager==4.0.1
Werkzeug==2.3.7
wsproto==1.2.0
